Spoolfool module compatible with new version interface.
I've changed the version check here, as it was buggy based on the original PR (seems like it worked by coincidence).
The MSRC article states that pretty much everything up to 21H2 is vulnerable. The PR tested against 21H2, so I'm confident
about the change. The ntdll revision number was checked independently of the version number, which I think is erroneous -
these reset with each new major release of Win10, so needs to be checked in coordination with specific build numbers.
So for now I've just removed it, since it was giving false positives and false negatives in different circumstances.